Update migrate function to be container aware
With this addition procs.py will generate the proc tree with
information about containers, e.g. A LXC process with have it's subtree
tagged as an LXC container. Default haul type is still
PID/non-container as before. RunC, Docker and OpenVZ support is not
assured yet but should be simple enough to add. Includes some
readability and error message improvements.